line between/after function scope, index.reducer.js rename to index.js, todo reducer options styling and fix some issues, pages start with capital letter
