+++ src/App.js
@@ -6,58 +6,40 @@ import { Button } from 'reactstrap';
 import TodoList from './components/TodoList';
 import { addTodo, updateTextArea, deleteTodo, toggleStatus } from './actions/todo.actions';
 
-class App extends Component {
+const App = ({ todos, deleteTodo, toggleStatus }) => {
+  const notFinishedTodosCount = todos.filter(todo => !todo.finished).length;
 
-  constructor(props) {
-    super(props);
-    this.onTodoToggleStatus = this.onTodoToggleStatus.bind(this);
-    this.onDeleteTodo = this.onDeleteTodo.bind(this);
-  }
-
-  onTodoToggleStatus(id) {
-    this.props.toggleStatus(id);
-  }
-
-  onDeleteTodo(id) {
-    this.props.deleteTodo(id);
-  }
-
-  render() {
-    const { todos } = this.props;
-    const notFinishedTodosCount = todos.filter(todo => !todo.finished).length;
-
-    return (
-      <div className="App-intro">
-        <div className="container">
-          <div className="row">
-            <div className="col-md-2"></div>
-            <div className="col-md-8">
-              <div className="todolist not-done">
-                <div>
-                  <h1>Todos</h1>
-                  <Link to="/new"><Button color="success" outline>Add New Todo</Button></Link>
-                  <hr />
-                  <TodoList
-                    onDeleteTodo={this.onDeleteTodo}
-                    onTextClick={this.onTodoToggleStatus}
-                    todos={todos}
-                  />
-                  <div
-                    className="todo-footer"
-                  >
-                    <strong>
-                      <span className="count-todos"></span>
-                    </strong>{`Items Left to do: ${notFinishedTodosCount}`}
-                  </div>
+  return (
+    <div className="App-intro">
+      <div className="container">
+        <div className="row">
+          <div className="col-md-2"></div>
+          <div className="col-md-8">
+            <div className="todolist not-done">
+              <div>
+                <h1>Todos</h1>
+                <Link to="/new"><Button color="success" outline>Add New Todo</Button></Link>
+                <hr />
+                <TodoList
+                  onDeleteTodo={deleteTodo}
+                  onTextClick={toggleStatus}
+                  todos={todos}
+                />
+                <div
+                  className="todo-footer"
+                >
+                  <strong>
+                    <span className="count-todos"></span>
+                  </strong>{`Items Left to do: ${notFinishedTodosCount}`}
                 </div>
               </div>
             </div>
           </div>
         </div>
       </div>
-    );
-  }
-}
+    </div>
+  );
+};
 
 function mapStateToProps(state, props) {
   return {
